Adding multisite upgrade suite
