+++ tests/ceph_ansible/test_ansible_upgrade.py
@@ -3,7 +3,7 @@ import logging
 import yaml
 
 from ceph.ceph_admin.common import config_dict_to_string
-from ceph.utils import get_ceph_versions, get_public_network
+from ceph.utils import get_ceph_versions, get_public_network, translate_to_ip
 from utility.utils import get_latest_container_image_tag
 
 log = logging.getLogger(__name__)
@@ -33,6 +33,13 @@ def run(ceph_cluster, **kw):
     ceph_cluster.custom_config = test_data.get("custom-config")
     ceph_cluster.custom_config_file = test_data.get("custom-config-file")
     ceph_cluster.use_cdn = config.get("use_cdn")
+
+    # Translate RGW node to ip address for Multisite
+    rgw_pull_host = config["ansi_config"].get("rgw_pullhost")
+    if rgw_pull_host:
+        ceph_cluster.ansible_config["rgw_pullhost"] = translate_to_ip(
+            kw["ceph_cluster_dict"], ceph_cluster.name, rgw_pull_host
+        )
 
     ceph_installer = ceph_cluster.get_ceph_object("installer")
     ansible_dir = "/usr/share/ceph-ansible"
