+++ tests/misc_env/exec.py
@@ -6,7 +6,7 @@ from typing import Any, Tuple
 
 from paramiko.channel import ChannelFile
 
-from ceph.utils import get_node_by_id
+from ceph.utils import get_node_by_id, translate_to_ip
 
 LOG = logging.getLogger(__name__)
 
@@ -33,36 +33,6 @@ def exec_command(node, **kwargs: Any) -> Tuple:
     out = out.read().decode() if isinstance(out, ChannelFile) else out
     err = err.read().decode() if isinstance(err, ChannelFile) else err
     return out, err
-
-
-def translate_to_ip(clusters, cluster_name: str, string: str) -> str:
-    """
-    Return the string after replacing ip: <node> pattern with the IP address of <node>.
-
-    In this method, the pattern {ip:<cluster>#<node>} would be replaced with the value
-    of node.ipaddress.
-
-    Args:
-        clusters:       Ceph cluster instance
-        cluster_name:   Name of the cluster under test.
-        string:         String that needs to be searched
-
-    Return:
-        String with node IDs replaced with IP addresses
-    """
-    replaced_string = string
-    node_list = re.findall("{node_ip:(.+?)}", string)
-
-    for node in node_list:
-        node_ = node
-        if "#" in node:
-            cluster_name, node = node.split("#")
-
-        node_ip = get_node_by_id(clusters[cluster_name], node).ip_address
-        replacement_pattern = "{node_ip:" + node_ + "}"
-        replaced_string = re.sub(replacement_pattern, node_ip, replaced_string)
-
-    return replaced_string
 
 
 def translate_to_hostname(cluster, string: str) -> str:
